Extract API class functionality into multiple modules
